+++ builds/wtf_wikipedia.cjs
@@ -1,4 +1,4 @@
-/*! wtf_wikipedia 10.0.5 MIT */
+/*! wtf_wikipedia 10.1.0 MIT */
 (function (global, factory) {
   typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory(require('isomorphic-unfetch')) :
   typeof define === 'function' && define.amd ? define(['isomorphic-unfetch'], factory) :
@@ -1533,8 +1533,8 @@
     nkcells: w + 'nkcells.info/index.php?title=$1',
     nara: 'catalog.archives.gov/id/$1',
     nosmoke: 'no-smok.net/nsmk/$1',
-    nost: 'nostalgia.' + wp,
-    nostalgia: 'nostalgia.' + wp,
+    nost: 'nostalgia' + wp,
+    nostalgia: 'nostalgia' + wp,
     oeis: 'oeis.org/$1',
     oldwikisource: 'wikisource.org/wiki/$1',
     olpc: 'wiki.laptop.org/go/$1',
@@ -1615,10 +1615,10 @@
     tabwiki: w + 'tabwiki.com/index.php/$1',
     tclerswiki: 'wiki.tcl.tk/$1',
     technorati: w + 'technorati.com/search/$1',
-    tenwiki: 'ten.' + wp,
-    testwiki: 'test.' + wp,
+    tenwiki: 'ten' + wp,
+    testwiki: 'test' + wp,
     testwikidata: 'test.wikidata.org/wiki/$1',
-    test2wiki: 'test2.' + wp,
+    test2wiki: 'test2' + wp,
     tfwiki: 'tfwiki.net/wiki/$1',
     thelemapedia: w + 'thelemapedia.org/index.php/$1',
     theopedia: w + 'theopedia.com/$1',
@@ -1651,7 +1651,7 @@
     voipinfo: w + 'voip-info.org/wiki/view/$1',
     votewiki: 'vote' + wm,
     werelate: w + 'werelate.org/wiki/$1',
-    wg: 'wg-en.' + wp,
+    wg: 'wg-en' + wp,
     wikia: w + 'wikia.com/wiki/w:c:$1',
     wikiasite: w + 'wikia.com/wiki/w:c:$1',
     wikiapiary: 'wikiapiary.com/wiki/$1',
@@ -1675,7 +1675,7 @@
     wikinvest: 'meta.wikimedia.org/wiki/Interwiki_map/discontinued#Wikinvest',
     wikiotics: 'wikiotics.org/$1',
     wikipapers: 'wikipapers.referata.com/wiki/$1',
-    wikipedia: 'en.' + wp,
+    wikipedia: 'en' + wp,
     wikipediawikipedia: 'en.wikipedia.org/wiki/Wikipedia:$1',
     wikiquote: 'en.wikiquote.org/wiki/$1',
     wikisophia: 'wikisophia.org/index.php?title=$1',
@@ -1776,23 +1776,23 @@
     zwiki: w + 'zwiki.org/$1',
     m: 'meta' + wm,
     meta: 'meta' + wm,
-    sep11: 'sep11.' + wp,
+    sep11: 'sep11' + wp,
     d: w + 'wikidata.org/wiki/$1',
-    minnan: 'zh-min-nan.' + wp,
-    nb: 'no.' + wp,
-    'zh-cfr': 'zh-min-nan.' + wp,
-    'zh-cn': 'zh.' + wp,
-    'zh-tw': 'zh.' + wp,
-    nan: 'zh-min-nan.' + wp,
-    vro: 'fiu-vro.' + wp,
-    cmn: 'zh.' + wp,
-    lzh: 'zh-classical.' + wp,
-    rup: 'roa-rup.' + wp,
-    gsw: 'als.' + wp,
-    'be-tarask': 'be-x-old.' + wp,
-    sgs: 'bat-smg.' + wp,
-    egl: 'eml.' + wp,
-    w: 'en.' + wp,
+    minnan: 'zh-min-nan' + wp,
+    nb: 'no' + wp,
+    'zh-cfr': 'zh-min-nan' + wp,
+    'zh-cn': 'zh' + wp,
+    'zh-tw': 'zh' + wp,
+    nan: 'zh-min-nan' + wp,
+    vro: 'fiu-vro' + wp,
+    cmn: 'zh' + wp,
+    lzh: 'zh-classical' + wp,
+    rup: 'roa-rup' + wp,
+    gsw: 'als' + wp,
+    'be-tarask': 'be-x-old' + wp,
+    sgs: 'bat-smg' + wp,
+    egl: 'eml' + wp,
+    w: 'en' + wp,
     wikt: 'en.wiktionary.org/wiki/$1',
     q: 'en.wikiquote.org/wiki/$1',
     b: 'en.wikibooks.org/wiki/$1',
@@ -1840,7 +1840,7 @@
   };
 
   const ignore_links =
-    /^:?(category|catégorie|kategorie|categoría|categoria|categorie|kategoria|تصنيف|image|file|fichier|datei|media):/i;
+    /^(category|catégorie|kategorie|categoría|categoria|categorie|kategoria|تصنيف|image|file|fichier|datei|media):/i;
   const external_link = /\[(https?|news|ftp|mailto|gopher|irc)(:\/\/[^\]| ]{4,1500})([| ].*?)?\]/g;
   const link_reg = /\[\[(.{0,1600}?)\]\]([a-z]+)?/gi; //allow dangling suffixes - "[[flanders]]s"
 
@@ -1912,6 +1912,10 @@
           obj.text = obj.page;
         }
         obj.page = obj.page;
+      }
+      // support [[:Category:Foo]] syntax
+      if (obj.text && obj.text.startsWith(':')) {
+        obj.text = obj.text.replace(/^:/, '');
       }
       links.push(obj);
       return s
@@ -3661,6 +3665,7 @@
   const isFile = new RegExp('(' + images.join('|') + '):', 'i');
   let fileNames = `(${images.join('|')})`;
   const file_reg = new RegExp(fileNames + ':(.+?)[\\||\\]]', 'iu');
+  const linkToFile = /^\[\[:/;
 
   //style directives for Wikipedia:Extended_image_syntax
   const imgLayouts = {
@@ -3684,6 +3689,9 @@
   const oneImage = function (img, doc) {
     let m = img.match(file_reg);
     if (m === null || !m[2]) {
+      return null
+    }
+    if (linkToFile.test(img)) {
       return null
     }
     let file = `${m[1]}:${m[2] || ''}`;
@@ -3727,8 +3735,8 @@
         let img = oneImage(s, doc);
         if (img) {
           paragraph.images.push(img);
+          wiki = wiki.replace(s, '');
         }
-        wiki = wiki.replace(s, '');
       }
     });
     paragraph.wiki = wiki;
@@ -5849,6 +5857,26 @@
       list.push(obj);
       return ''
     },
+    //https://en.wikipedia.org/wiki/Template:MedalCount
+    'medalcount': (tmpl, list) => {
+      let all = parser(tmpl).list || [];
+      let lines = [];
+      for (let i = 0; i < all.length; i += 4) {
+        lines.push({
+          name: all[i],
+          '1st': Number(all[i + 1]),
+          '2nd': Number(all[i + 2]),
+          '3rd': Number(all[i + 3]),
+        });
+        console.log(all[i]);
+      }
+      let obj = {
+        template: 'medalcount',
+        list: lines
+      };
+      list.push(obj);
+      return ''
+    }
   };
 
   let templates$4 = {
@@ -7687,7 +7715,7 @@
         return
       }
       // is it an Infobox?
-      if (json.template === 'infobox' && json.subbox !== 'yes' && !obj.nested) {
+      if (json.template === 'infobox' && json.subbox !== 'yes') {
         json.domain = domain; //infoboxes need this for images, i guess
         json.data = json.data || {}; //validate it a little
         res.infoboxes.push(new Infobox(json, obj.wiki));
@@ -8750,7 +8778,7 @@
     return removeReferenceSection(sections)
   };
 
-  const cat_reg = new RegExp('\\[\\[:?(' + _categories.join('|') + '):(.{2,178}?)]](w{0,10})', 'gi');
+  const cat_reg = new RegExp('\\[\\[(' + _categories.join('|') + '):(.{2,178}?)]](w{0,10})', 'gi');
   const cat_remove_reg = new RegExp('^\\[\\[:?(' + _categories.join('|') + '):', 'gi');
 
   const parse_categories = function (wiki) {
@@ -9437,7 +9465,7 @@
       })
   };
 
-  var version = '10.0.5';
+  var version = '10.1.0';
 
   /**
    * use the native client-side fetch function